+++ internal/models/appfeedback.go
@@ -9,7 +9,7 @@ import (
 type AppFeedback struct {
 	cm.BaseModel
 
-	Rating   int       `json:"rating" db:"rating"`
+	Rating   float64   `json:"rating" db:"rating"`
 	Feedback string    `json:"feedback" db:"feedback"`
 	SubmitAt time.Time `json:"submit_at" db:"submit_at"`
 	UserID   int       `json:"user_id" db:"user_id"`
